refactor(ts): extract timeline/shifts.ts from store
Move the time-domain action manipulation (refreshAllActionShifts, getShiftedEndTime,
ultimate-enhancement metrics, pushSubsequentActions/pullSubsequentActions,
toGameTime/toRealTime) into a useShifts(deps) composable, mirroring the optimizer.
Relocates UltEnhancer/UltEnhancerContext to timeline/types.ts. Injects the shared
action-kind predicates and ult-enhancement extender lookup as deps. No behavior change.
