+++ src/stores/timeline/types.ts
@@ -60,6 +60,18 @@ export interface TrackStatusState {
   [key: string]: unknown;
 }
 
+/** Context passed to an ultimate-enhancement extender to compute extra freeze duration. */
+export interface UltEnhancerContext {
+  track: Track | null | undefined;
+  enhStart: number;
+  baseDuration: number;
+  ultimateAction: TimelineAction;
+  getShiftedEndTime: (start: number, duration: number, instanceId?: string) => number;
+}
+
+/** An operator-specific hook that extends an ultimate's enhancement window. */
+export type UltEnhancer = (context: UltEnhancerContext) => number;
+
 export interface TriggerEffectEntry {
   triggerEffect?: unknown;
   sourceTrackId?: string;
